+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Added
 ### Changed
 ### Fixed
+- For a Label with a hidden field, apply the field's id to the hidden field, not the label div
+- Do not raise an exception if a Select has a selected value that is not included in options or disabled options
 
 ## [2.1.0] - 2023-050-29
 ### Added