+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Added
 ### Changed
 - Expand/Collapse can be passed a `parent_dom_id` to be used as the parent node of the fold ups instead of the default (form)
+- DateTime control can optionally accept a `default_hour` and `default_minute` to set the time even if the datetime is nil
 ### Fixed
 - For a Label with a hidden field, apply the field's id to the hidden field, not the label div
 - Do not raise an exception if a Select has a selected value that is not included in options or disabled options